# backend/docs/openapi.json
{
    "components": {
        "schemas": {
...
            "auth.LoginRequest": {
                "properties": {
                    "password": {
                        "example": "ClaveSegura123",
                        "type": "string"
                    },
                    "username": {
                        "example": "estudiante1",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "auth.LoginResponse": {
                "properties": {
                    "message": {
                        "example": "Autenticación exitosa",
                        "type": "string"
                    },
                    "token": {
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "type": "string"
                    },
                    "username": {
                        "example": "estudiante1",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "auth.MeResponse": {
                "properties": {
                    "role": {
                        "example": "student",
                        "type": "string"
                    },
                    "username": {
                        "example": "estudiante1",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "auth.ResendVerificationRequest": {
                "properties": {
                    "email": {
                        "example": "estudiante1@ejemplo.com",
                        "type": "string"
                    },
                    "username": {
                        "example": "estudiante1",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "utils.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Se requieren username, email y password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "utils.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "Operación exitosa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Token JWT de sesión. Formato: `Bearer \u003ctoken\u003e`.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "API REST del monolito modular: identidad y sesiones, administración y auditoría, autoría de cursos versionados y carga multimedia con procesamiento asíncrono a HLS.\nAutenticación con JWT revocable: iniciar sesión en /api/v1/auth/login y enviar `Authorization: Bearer \u003ctoken\u003e`.",
        "title": "Plataforma MOOC - API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "description": "Verifica las credenciales y retorna un token JWT de 24 horas para el header Authorization.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.LoginRequest",
                                "summary": "request",
                                "description": "Credenciales"
                            }
                        }
                    },
                    "description": "Credenciales",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.LoginResponse"
                                }
                            }
                        },
                        "description": "Autenticación exitosa"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Credenciales incorrectas"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Cuenta desactivada o bloqueada"
                    }
                },
                "summary": "Iniciar sesión",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "description": "Invalida el token de la petición actual. Es idempotente.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Sesión cerrada"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Token inválido o ausente"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Cerrar sesión",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "description": "Retorna el username y rol asociados al token de la petición.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.MeResponse"
                                }
                            }
                        },
                        "description": "Perfil de la sesión"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Token inválido o ausente"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Perfil de la sesión",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "Registra un estudiante o administrador de forma pública. Crear un profesor requiere el token de un administrador.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e (obligatorio solo para rol professor)",
                        "in": "header",
                        "name": "Authorization",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.RegisterRequest",
                                "summary": "request",
                                "description": "Datos de registro"
                            }
                        }
                    },
                    "description": "Datos de registro",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Usuario registrado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida o rol inválido"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Token de administrador inválido"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Se requiere administrador para crear profesores"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Usuario o correo ya en uso"
                    }
                },
                "summary": "Registrar usuario",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/resend-verification": {
            "post": {
                "description": "Genera un nuevo código de verificación y lo envía al correo indicado. Solo para cuentas pendientes de verificación.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.ResendVerificationRequest",
                                "summary": "request",
                                "description": "Usuario y correo"
                            }
                        }
                    },
                    "description": "Usuario y correo",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Código reenviado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Usuario no encontrado"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Usuario ya verificado"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No se pudo enviar el correo"
                    }
                },
                "summary": "Reenviar verificación",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "description": "Valida el código de recuperación y guarda la nueva contraseña (hasheada con bcrypt).",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.ResetPasswordRequest",
                                "summary": "request",
                                "description": "Usuario, código y nueva contraseña"
                            }
                        }
                    },
                    "description": "Usuario, código y nueva contraseña",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Contraseña restablecida"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida"
                    }
                },
                "summary": "Restablecer contraseña",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/revoke-session": {
            "post": {
                "description": "Elimina de Redis el token indicado en el cuerpo o en el header Authorization. No requiere autenticación (uso en pruebas).",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/auth.TokenRequest",
                                "summary": "request",
                                "description": "Token a revocar (o header Authorization)"
                            }
                        }
                    },
                    "description": "Token a revocar (o header Authorization)"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Sesión revocada"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Falta el token a revocar"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No se pudo revocar la sesión"
                    }
                },
                "summary": "Revocar una sesión",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/send-recovery-code": {
            "get": {
                "description": "Genera un código de recuperación válido por 10 minutos y lo envía al correo del usuario.",
                "parameters": [
                    {
                        "description": "Nombre de usuario",
                        "in": "query",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Correo (informativo)",
                        "in": "query",
                        "name": "email",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Código enviado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida"
                    }
                },
                "summary": "Solicitar recuperación",
                "tags": [
                    "Autenticación"
                ]
            }
        },
        "/api/v1/auth/verify": {
            "get": {
                "description": "Activa la cuenta recién registrada con el código de verificación de 6 caracteres.",
                "parameters": [
                    {
                        "description": "Nombre de usuario",
                        "in": "query",
                        "name": "username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Código de verificación",
                        "in": "query",
                        "name": "code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Usuario verificado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Solicitud inválida o código incorrecto"
                    }
                },
                "summary": "Verificar cuenta",
                "tags": [
                    "Autenticación"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080"
        }
    ]
}